Added act subscriptions for care plan service, added auto-merge capability

+++ OpenIZ.Mobile.Core/Protocol/CarePlanManagerService.cs
@@ -31,6 +31,7 @@ using System;
 using System.Collections.Generic;
 using System.Dynamic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
@@ -49,6 +50,9 @@ namespace OpenIZ.Mobile.Core.Protocol
 
         // Tracer
         private Tracer m_tracer = Tracer.GetTracer(typeof(CarePlanManagerService));
+
+        // Represents a promise to perform a care plan
+        private HashSet<Guid> m_actCarePlanPromise = new HashSet<Guid>();
 
         /// <summary>
         /// True when running
@@ -130,10 +134,14 @@ namespace OpenIZ.Mobile.Core.Protocol
                         {
                             foreach (var i in e.Data.Item.OfType<Patient>())
                                 this.UpdateCarePlan(i);
+                            foreach (var i in e.Data.Item.OfType<Act>())
+                                this.UpdateCarePlan(i);
                         };
                         bundlePersistence.Updated += (o, e) =>
                         {
                             foreach (var i in e.Data.Item.OfType<Patient>())
+                                this.UpdateCarePlan(i);
+                            foreach (var i in e.Data.Item.OfType<Act>())
                                 this.UpdateCarePlan(i);
                         };
                         bundlePersistence.Obsoleted += (o, e) =>
@@ -142,17 +150,135 @@ namespace OpenIZ.Mobile.Core.Protocol
                             var dataMart = warehouseService.GetDatamart("oizcp");
                             foreach (var i in e.Data.Item.OfType<Patient>())
                                 warehouseService.Delete(dataMart.Id, new { patient_id = i.Key.Value });
+                            foreach (var i in e.Data.Item.OfType<Act>())
+                                this.UpdateCarePlan(i);
                         };
 
                     }
+
+                    // Act persistence services
+                    foreach (var t in typeof(Act).GetTypeInfo().Assembly.ExportedTypes.Where(o => o == typeof(Act) || typeof(Act).GetTypeInfo().IsAssignableFrom(o.GetTypeInfo()) && !o.GetTypeInfo().IsAbstract))
+                    {
+                        var pType = typeof(IDataPersistenceService<>).MakeGenericType(t);
+                        var pInstance = ApplicationContext.Current.GetService(pType) as IDataPersistenceService;
+
+                        // Create a delegate which calls UpdateCarePlan
+                        // Construct the delegate
+                        var ppeArgType = typeof(DataPersistenceEventArgs<>).MakeGenericType(t);
+                        var evtHdlrType = typeof(EventHandler<>).MakeGenericType(ppeArgType);
+                        var senderParm = Expression.Parameter(typeof(Object), "o");
+                        var eventParm = Expression.Parameter(ppeArgType, "e");
+                        var eventData = Expression.Convert(Expression.MakeMemberAccess(eventParm, ppeArgType.GetRuntimeProperty("Data")), typeof(Act));
+                        var insertInstanceDelegate = Expression.Lambda(evtHdlrType, Expression.Call(Expression.Constant(this), typeof(CarePlanManagerService).GetRuntimeMethod(nameof(UpdateCarePlan), new Type[] { typeof(Act) }), eventData), senderParm, eventParm).Compile();
+                        var updateInstanceDelegate = Expression.Lambda(evtHdlrType, Expression.Call(Expression.Constant(this), typeof(CarePlanManagerService).GetRuntimeMethod(nameof(UpdateCarePlan), new Type[] { typeof(Act) }), eventData), senderParm, eventParm).Compile();
+                        var obsoleteInstanceDelegate = Expression.Lambda(evtHdlrType, Expression.Call(Expression.Constant(this), typeof(CarePlanManagerService).GetRuntimeMethod(nameof(UpdateCarePlan), new Type[] { typeof(Act) }), eventData), senderParm, eventParm).Compile();
+
+                        // Bind to events
+                        pType.GetRuntimeEvent("Inserted").AddEventHandler(pInstance, insertInstanceDelegate);
+                        pType.GetRuntimeEvent("Updated").AddEventHandler(pInstance, updateInstanceDelegate);
+                        pType.GetRuntimeEvent("Obsoleted").AddEventHandler(pInstance, obsoleteInstanceDelegate);
+
+                    }
+
                 }
-                catch(Exception e)
+                catch (Exception e)
                 {
                     this.m_tracer.TraceError("Could not bind clinical protocols: {0}", e);
                 }
             };
             this.Started?.Invoke(this, EventArgs.Empty);
             return true;
+        }
+
+        /// <summary>
+        /// Update the care plan given that a new act exists
+        /// </summary>
+        public void UpdateCarePlan(Act a)
+        {
+            ApplicationContext.Current.GetService<IThreadPoolService>().QueueUserWorkItem((d) =>
+            {
+                var act = d as Act;
+                // First step, we delete all acts in the warehouse for the specified patient in the protocol
+                var patientId = act.Participations.FirstOrDefault(o => o.ParticipationRoleKey == ActParticipationKey.RecordTarget)?.PlayerEntityKey;
+                if (patientId == null)
+                {
+                    this.m_tracer.TraceError("Cannot update care plan for act as it seems to have no RecordTarget");
+                    return;
+                }
+
+                try
+                {
+
+                    IEnumerable<Act> carePlan = null;
+                    var careplanService = ApplicationContext.Current.GetService<ICarePlanService>();
+                    var warehouseService = ApplicationContext.Current.GetService<IAdHocDatawarehouseService>();
+                    var dataMart = warehouseService.GetDatamart("oizcp");
+                    var patient = ApplicationContext.Current.GetService<IDataPersistenceService<Patient>>().Get(patientId.Value).Clone() as Patient;
+                    patient.Participations = new List<ActParticipation>(patient.Participations);
+
+                    // Is there a protocol for this act?
+                    if (act.Protocols.Count() == 0)
+                    {
+                        if (this.m_actCarePlanPromise.Contains(patient.Key.Value)) return;
+                        else lock (this.m_actCarePlanPromise)
+                            {
+                                if (this.m_actCarePlanPromise.Contains(patient.Key.Value)) return;
+                                else this.m_actCarePlanPromise.Add(patient.Key.Value);
+                            }
+
+                        // Need to re-calculate the entire care plan
+                        this.m_tracer.TraceWarning("Will need to calculate the entire care plan for patient {0}", patientId);
+
+                        this.m_tracer.TraceVerbose("Calculating care plan for {0}", patient.Key);
+
+                        // First, we clear the warehouse
+                        warehouseService.Delete(dataMart.Id, new { patient_id = patient.Key.Value });
+
+                        // Now calculate
+                        carePlan = careplanService.CreateCarePlan(patient, false);
+                    }
+                    else
+                    {
+                        warehouseService.Delete(dataMart.Id, new
+                        {
+                            patient_id = patientId.Value,
+                            protocol_id = act.Protocols.FirstOrDefault()?.ProtocolKey
+                        });
+
+                        carePlan = careplanService.CreateCarePlan(patient, false, act.Protocols.FirstOrDefault().ProtocolKey);
+                    }
+
+
+                    /// Create a plan for the warehouse
+                    var warehousePlan = carePlan.Select(o => new
+                    {
+                        creation_date = DateTime.Now,
+                        patient_id = patient.Key.Value,
+                        location_id = patient.Relationships.FirstOrDefault(r => r.RelationshipTypeKey == EntityRelationshipTypeKeys.DedicatedServiceDeliveryLocation || r.RelationshipType?.Mnemonic == "DedicatedServiceDeliveryLocation")?.TargetEntityKey.Value,
+                        act_id = o.Key.Value,
+                        protocol_id = o.Protocols.FirstOrDefault().ProtocolKey,
+                        class_id = o.ClassConceptKey.Value,
+                        type_id = o.TypeConceptKey.Value,
+                        min_date = o.StartTime?.DateTime,
+                        max_date = o.StopTime?.DateTime,
+                        act_date = o.ActTime.DateTime,
+                        product_id = o.Participations?.FirstOrDefault(r => r.ParticipationRoleKey == ActParticipationKey.Product || r.ParticipationRole?.Mnemonic == "Product")?.PlayerEntityKey.Value
+                    });
+
+                    // Insert plans
+                    warehouseService.Add(dataMart.Id, warehousePlan);
+                }
+                catch (Exception ex)
+                {
+                    this.m_tracer.TraceError("Could not update care plan based on Act {0}: {1}", act, ex);
+                }
+                finally
+                {
+                    lock (this.m_actCarePlanPromise)
+                        this.m_actCarePlanPromise.Remove(patientId.Value);
+                }
+            }, a.Clone());
+
         }
 
         /// <summary>
@@ -164,34 +290,55 @@ namespace OpenIZ.Mobile.Core.Protocol
             {
                 var data = d as Patient;
 
-                data.Participations = new List<ActParticipation>(data.Participations);
-                this.m_tracer.TraceVerbose("Calculating care plan for {0}", data.Key);
-
-                // First, we clear the warehouse
-                var warehouseService = ApplicationContext.Current.GetService<IAdHocDatawarehouseService>();
-                var dataMart = warehouseService.GetDatamart("oizcp");
-                warehouseService.Delete(dataMart.Id, new { patient_id = data.Key.Value });
-                var careplanService = ApplicationContext.Current.GetService<ICarePlanService>();
-
-                // Now calculate
-                var carePlan = careplanService.CreateCarePlan(data, false);
-                var warehousePlan = carePlan.Select(o => new
+                try
                 {
-                    creation_date = DateTime.Now,
-                    patient_id = data.Key.Value,
-                    location_id = data.Relationships.FirstOrDefault(r => r.RelationshipTypeKey == EntityRelationshipTypeKeys.DedicatedServiceDeliveryLocation || r.RelationshipType?.Mnemonic == "DedicatedServiceDeliveryLocation")?.TargetEntityKey.Value,
-                    act_id = o.Key.Value,
-                    class_id = o.ClassConceptKey.Value,
-                    type_id = o.TypeConceptKey.Value,
-                    protocol_id = o.Protocols.FirstOrDefault()?.ProtocolKey,
-                    min_date = o.StartTime?.DateTime,
-                    max_date = o.StopTime?.DateTime,
-                    act_date = o.ActTime.DateTime,
-                    product_id = o.Participations?.FirstOrDefault(r => r.ParticipationRoleKey == ActParticipationKey.Product || r.ParticipationRole?.Mnemonic == "Product")?.PlayerEntityKey.Value
-                });
+                    if (this.m_actCarePlanPromise.Contains(data.Key.Value)) return;
+                    else lock (this.m_actCarePlanPromise)
+                        {
+                            if (this.m_actCarePlanPromise.Contains(data.Key.Value)) return;
+                            else this.m_actCarePlanPromise.Add(data.Key.Value);
+                        }
 
-                // Insert plans
-                warehouseService.Add(dataMart.Id, warehousePlan);
+                    data.Participations = new List<ActParticipation>(data.Participations);
+                    this.m_tracer.TraceVerbose("Calculating care plan for {0}", data.Key);
+
+                    // First, we clear the warehouse
+                    var warehouseService = ApplicationContext.Current.GetService<IAdHocDatawarehouseService>();
+                    var dataMart = warehouseService.GetDatamart("oizcp");
+                    warehouseService.Delete(dataMart.Id, new { patient_id = data.Key.Value });
+                    var careplanService = ApplicationContext.Current.GetService<ICarePlanService>();
+
+                    // Now calculate
+                    var carePlan = careplanService.CreateCarePlan(data, false);
+                    var warehousePlan = carePlan.Select(o => new
+                    {
+                        creation_date = DateTime.Now,
+                        patient_id = data.Key.Value,
+                        location_id = data.Relationships.FirstOrDefault(r => r.RelationshipTypeKey == EntityRelationshipTypeKeys.DedicatedServiceDeliveryLocation || r.RelationshipType?.Mnemonic == "DedicatedServiceDeliveryLocation")?.TargetEntityKey.Value,
+                        act_id = o.Key.Value,
+                        class_id = o.ClassConceptKey.Value,
+                        type_id = o.TypeConceptKey.Value,
+                        protocol_id = o.Protocols.FirstOrDefault()?.ProtocolKey,
+                        min_date = o.StartTime?.DateTime,
+                        max_date = o.StopTime?.DateTime,
+                        act_date = o.ActTime.DateTime,
+                        product_id = o.Participations?.FirstOrDefault(r => r.ParticipationRoleKey == ActParticipationKey.Product || r.ParticipationRole?.Mnemonic == "Product")?.PlayerEntityKey.Value
+                    });
+
+                    // Insert plans
+                    warehouseService.Add(dataMart.Id, warehousePlan);
+
+                }
+                catch (Exception ex)
+                {
+                    this.m_tracer.TraceError("Could not update care plan for Patient {0}: {1}", p, ex);
+                }
+                finally
+                {
+                    lock (this.m_actCarePlanPromise)
+                        this.m_actCarePlanPromise.Remove(data.Key.Value);
+                }
+
             }, p.Clone());
         }
